requirements.txt
moviepy==1.0.3
openai-whisper==20231117
torch
torchvision
torchaudio
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
pexels-api>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pydub>=0.25.1
scikit-image>=0.21.0
